Colocate invite dialog handoff in InviteFriendsAnnouncement instead of NUX context
